Refactored some classes moving them to new files and created wrapper sub-class for GAE torch-geometric class, implementing forward() method, for QOL purposes.
